Split amp-instagram component files
