+++ extensions/amp-instagram/1.0/component.js
